Add experimental flag to use suspend instad of freeze for Xen VM suspend
When suspending a VM the Xen driver currently uses the "freeze" methods
for going to sleep. This doesn't put devices in low power state, at
least for some drivers (for example USB). But for S0ix we need to put
all (real) devices into their low power state. For this the "suspend"
methods needs to be used.

(See Documentation/driver-api/pm/devices.rst for description of those
methods.)

To be able to include this in a stable release put this behind an
experimental flag such that by default the old behavior is unmodified.
This flag is unstable and likely will go away.

TODO: If this should be upstreamed it needs to be clarified when the old
behavior in xenbus_frontend_dev_resume is needed (save/restore, I guess)
and the new behavior adapted accordingly.